refactor: UserRepository 상위 폴더로 이동
