+++ src/api/UserRepository.ts
@@ -1,4 +1,4 @@
-import apiClient from "../apiClient";
+import apiClient from "./apiClient";
 
 class UserRepository {
